.history/backend/requirements_20250926115330.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
SQLAlchemy
alembic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
sentence-transformers
scikit-learn
spacy
pypdf                # <-- add (PDF text)
python-docx    


